split trace and traceDB / add trace loading without folder
