+++ .gitignore
@@ -3,5 +3,6 @@
 *debug_log*
 **.idea**
 **.txt
-**raw_traces_cache**
+**raw_traces**
 **pics**
+
